@@ -1,18 +1,28 @@
 require('babel-register')();
 
-var jsdom = require('jsdom').jsdom;
+const { JSDOM } = require('jsdom');
 
-var exposedProperties = ['window', 'navigator', 'document'];
+const jsdom = new JSDOM('<!doctype html><html><body></body></html>');
+const { window } = jsdom;
 
-global.document = jsdom('');
-global.window = document.defaultView;
-Object.keys(document.defaultView).forEach((property) => {
-  if (typeof global[property] === 'undefined') {
-    exposedProperties.push(property);
-    global[property] = document.defaultView[property];
-  }
-});
+const { configure } = require('enzyme');
+const Adapter = require('enzyme-adapter-react-16');
 
+configure({ adapter: new Adapter() });
+
+function copyProps(src, target) {
+  const props = Object.getOwnPropertyNames(src)
+    .filter(prop => typeof target[prop] === 'undefined')
+    .reduce((result, prop) => ({
+      ...result,
+      [prop]: Object.getOwnPropertyDescriptor(src, prop),
+    }), {});
+  Object.defineProperties(target, props);
+}
+
+global.window = window;
+global.document = window.document;
 global.navigator = {
-  userAgent: 'node.js'
+  userAgent: 'node.js',
 };
+copyProps(window, global);
